+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The changelog of the TexLab language server can be found [here](https://github.c
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Current
+
+### Changed
+
+- Add the `texlab.findEnvironments` command
+- Add the `texlab.build.filename` setting
+
 ## [5.21.0] - 2024-10-26
 
 ### Changed